Simplify code in main (move under home_page)
